feat: Split user/validator layout
