+++ src/app/(main)/activities/[id]/page.tsx
