+++ src/app/(main)/activities/page.tsx
