+++ src/app/(main)/layout/index.tsx
@@ -1,0 +1,23 @@
+import Header from './Header';
+import Footer from './Footer';
+
+export const metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+};
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const RootLayout = ({ children }: Props) => {
+  return (
+    <>
+      <Header />
+      <div style={{ minHeight: 'calc(100vh - 379px)' }}>{children}</div>
+      <Footer />
+    </>
+  );
+};
+
+export default RootLayout;
