+++ src/app/(main)/purchasing-process/(reservation)/select-area/page.tsx
