--- src/app/admin/activities/[id]/page.tsx
@@ -1,5 +1,0 @@
-const Id = () => {
-  return <div>Id</div>;
-};
-
-export default Id;
