--- src/app/admin/activities/new/page.tsx
@@ -1,5 +1,0 @@
-const New = () => {
-  return <div>New</div>;
-};
-
-export default New;
