--- src/app/admin/activities/page.tsx
@@ -1,5 +1,0 @@
-const Nes = () => {
-  return <div>Nes</div>;
-};
-
-export default Nes;
