--- src/app/admin/event/page.tsx
@@ -1,5 +1,0 @@
-const Event = () => {
-  return <div>Event</div>;
-};
-
-export default Event;
