--- src/app/admin/event/scan/page.tsx
@@ -1,5 +1,0 @@
-const Scan = () => {
-  return <div>Scan</div>;
-};
-
-export default Scan;
