--- src/app/admin/info/page.tsx
@@ -1,5 +1,0 @@
-const Info = () => {
-  return <div>Info</div>;
-};
-
-export default Info;
