+++ src/app/layout/index.tsx
@@ -1,30 +1,18 @@
-import Header from './Header';
-import Footer from './Footer';
-import Providers from './providers';
+import Providers from '../providers';
 import Dialog from './Dialog';
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: 'Next.js',
+  description: 'Generated by Next.js',
 };
 
-type Props = {
-  children: React.ReactNode;
-};
-
-const RootLayout = ({ children }: Props) => {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="zh-TW">
       <body>
-        <Providers>
-          <Dialog />
-          <Header />
-          <div style={{ minHeight: 'calc(100vh - 379px)' }}>{children}</div>
-          <Footer />
-        </Providers>
+        <Dialog />
+        <Providers>{children}</Providers>
       </body>
     </html>
   );
-};
-
-export default RootLayout;
+}
